core: tests: stdcm: split tests into several new files
